Update debugger.php
It's a simplification. I've been using this class and this way is more simple and fast for find bugs.
In this branch you can use only the constructor with parameter and it's done. (<?php $debugger= new Debugger('Test',$array,true); ?>)
But in this new branch the pathFile is stantard. If you want more than one JSON, you may use the first branch.